Adding `-h` support, passing args to `git log`
Showing a help message on `-h` or `--help` is always good. (Note: if launched through `git author-stats --help`, `git` itself will try to open the manpage, which will fail anyway.)

Also passing the arguments to `git log`, which allows some customization, such as restricting for a sub path or restricting a date range.